+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - New plugin `on_demand` to implement page rendering on demand.
 - New option `watcher` to configure the watcher of the live reload
   (the debounce interval and a list of ignored paths).
+- Improved default 404 page of the local server.
 
 ### Changed
 - BREAKING: Some internal parts have been refactored for more flexibility. This change affects to Lume types in `core.ts`.